refactor: use autoload for platform-specific implementations
- Extract InstallDeps::ForMacOS and ForLinux to separate files
- Use autoload for ForMacOS/ForLinux in both Packing and InstallDeps
- Platform-specific code is only loaded when needed
